force overwrite of ~/.config/mimeapps.list file
Programs make changes to this file, which then forces it to be backed
up, but home-manager will refuse to back up if a backup exists, instead
of simply overwriting like NixOS does.

As only this file seems affected currently, force overwrite of it.

See https://github.com/nix-community/home-manager/issues/4199 for more
context.